fixes
Former-commit-id: 2639e7e12b77aabd130e7d8a7a9fb39effbfe280 [formerly b531cb9ea9a3e7b891a459d5e7bfa79e277d65a3 [formerly 28f0b472b6e77a9d193320a258dd5d61de542c94]]
Former-commit-id: a3a773e7a7d1e3893dd7ec750b4ec703cb484ec4
Former-commit-id: c406484cec279f0842a76277a6ae99de4b46a2e3

+++ artap/datastore.py
@@ -376,7 +376,8 @@ class FileDataStore(DataStore):
 
         self.database_name = database_name
         if remove_existing:
-            os.remove(self.database_name)
+            if os._exists(self.database_name):
+                os.remove(self.database_name)
 
         # self.db = SqliteDict(self.db_fn, autocommit=True)
         # self.db = SqliteDict(":memory", autocommit=False)
